requirements.txt
django-ninja
Django
ipython
pytest
pytest-django
ruff
model-bakery
uvicorn
psycopg
